+++ app/components/ram.py
@@ -1,7 +1,7 @@
 import psutil
 
 
-def get_ram_data():
+def get_ram_data() -> dict:
     ram = psutil.virtual_memory()
     return {
         "total": ram.total,